+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/Charges/IChargeMessageRepository.cs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace GreenEnergyHub.Charges.Domain.Charges
@@ -29,12 +27,5 @@ namespace GreenEnergyHub.Charges.Domain.Charges
         /// <param name="chargeMessage"></param>
         /// <returns>A <see cref="Task"/>.</returns>
         Task AddAsync(ChargeMessage chargeMessage);
-
-        /// <summary>
-        /// Get ChargeMessages by charge id.
-        /// </summary>
-        /// <param name="chargeId"></param>
-        /// <returns>A readonly collection of ChargeMessages.</returns>
-        Task<IReadOnlyCollection<ChargeMessage>> GetByChargeIdAsync(Guid chargeId);
     }
 }