+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.TestCore/Builders/Command/DocumentDtoBuilder.cs
@@ -12,7 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System;
 using GreenEnergyHub.Charges.Domain.Dtos.SharedDtos;
+using GreenEnergyHub.Charges.TestCore.Data;
 using NodaTime;
 
 namespace GreenEnergyHub.Charges.TestCore.Builders.Command
@@ -30,7 +32,7 @@ namespace GreenEnergyHub.Charges.TestCore.Builders.Command
 
         public DocumentDtoBuilder()
         {
-            _id = "id";
+            _id = Guid.NewGuid().ToString("N");
             _recipient = new MarketParticipantDtoBuilder()
                 .WithMarketParticipantId(SeededData.MarketParticipants.MeteringPointAdministrator.Gln)
                 .WithMarketParticipantRole(MarketParticipantRole.MeteringPointAdministrator)